MASSIVE CHANGE!
influx_pre_header is now removed in favour for influx_device. This so we can tag in the database. That means if you go to this version all your data will end up differently!

Please check your database with "show series" before and after doing the change. This will force you to better utilize Influx systems.

More to come!